Move the wrapped types out of provider_interfaces
