+++ onnxruntime/core/providers/shared_library/provider_api.h
@@ -213,6 +213,7 @@ using NameMLValMap = std::unordered_map<std::string, OrtValue>;
 #include "core/framework/data_transfer.h"
 #include "core/framework/execution_provider.h"
 #include "provider_interfaces.h"
+#include "provider_wrappedtypes.h"
 #include "core/framework/op_kernel.h"
 #include "core/framework/data_types_internal.h"
 #include "core/framework/tensorprotoutils.h"
